During the 6 month Internship in Hahsedin by Deloitte I worked on two projects. Both of these projects were created from scratch. 1. MFPA : Mutual Fund Portfolio Analyzer. During the initial days in the Hashedin new interns are called as Linkers. They go through a rigorous traning from basic to advance over a span of 5 weeks. In the first week, we were given challenge to create few webpages using HTML CSS and git basics. 2nd week: A project in React 3rd week: A project in Spring boot 4th week: A project in Python Django 5th week: A complete end-to-end project with either Python Django-React or Spring Boot- React. During these 5 weeks we were working individually and trying to survive each week.
For the remaining days in the 2 months journey of the Hashed University
I was put in a team with other linkers to create a project and compete
with other teams. Out of many project ideas we were given a project to
demonstrate how can we create a Mutual Fund Portfolio Analyzer using
Mulesoft.

I was working as a backend developer and worked on both Mulesoft and
Spring boot application.

After HU, we became Hashers.

The first project I got after becoming Hasher was HBS. We created this
product from scratch. It was team of four developers. There were one
backend developer, one frontend developer, one lead developer, working
as a front end developer and leading discussions with Product Manager
and team.
We were able to successfully create this project and host on Google
Store.
A complete usable product without any bugs.